requirements.txt
aiogram
Pyrogram
python-dotenv
redis
Telethon
tgcrypto
